Revert "Update main.c"
This reverts commit f6ea11931a863b6d88d1431daf0c5a4dd84370b8.

+++ src/main.c
@@ -168,7 +168,7 @@ int main(void) {
 
   // Save bootloader version to pre-defined register, retrieved by application
   // TODO move to CF2
-  BOOTLOADER_VERSION_REGISTER = (uint32_t)(MK_BOOTLOADER_VERSION);
+  BOOTLOADER_VERSION_REGISTER = (MK_BOOTLOADER_VERSION);
 
   board_init();
   bootloader_init();
